# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "AdjustPartBody": {
        "properties": {
          "adj_type": {
            "enum": [
              "set",
              "add",
              "remove"
            ],
            "title": "Adj Type",
            "type": "string"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "source": {
            "default": "",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "adj_type",
          "quantity"
        ],
        "title": "AdjustPartBody",
        "type": "object"
      },
...
      "ConsumeBomBody": {
        "properties": {
          "board_qty": {
            "title": "Board Qty",
            "type": "integer"
          },
          "bom_name": {
            "title": "Bom Name",
            "type": "string"
          },
          "matches": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Matches",
            "type": "array"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "source": {
            "default": "",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "matches",
          "board_qty",
          "bom_name"
        ],
        "title": "ConsumeBomBody",
        "type": "object"
      },
...
      "ExtractSpecFromValueBody": {
        "properties": {
          "package_str": {
            "title": "Package Str",
            "type": "string"
          },
          "part_type": {
            "title": "Part Type",
            "type": "string"
          },
          "value_str": {
            "title": "Value Str",
            "type": "string"
          }
        },
        "required": [
          "part_type",
          "value_str",
          "package_str"
        ],
        "title": "ExtractSpecFromValueBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImportPurchasesBody": {
        "properties": {
          "rows": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "rows"
        ],
        "title": "ImportPurchasesBody",
        "type": "object"
      },
...
      "RecordFetchedPricesBody": {
        "properties": {
          "distributor": {
            "title": "Distributor",
            "type": "string"
          },
          "price_tiers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Price Tiers",
            "type": "array"
          }
        },
        "required": [
          "distributor",
          "price_tiers"
        ],
        "title": "RecordFetchedPricesBody",
        "type": "object"
      },
      "ResolveBomSpecBody": {
        "properties": {
          "package": {
            "title": "Package",
            "type": "string"
          },
          "part_type": {
            "title": "Part Type",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "part_type",
          "value",
          "package"
        ],
        "title": "ResolveBomSpecBody",
        "type": "object"
      },
...
      "UpdatePartFieldsBody": {
        "properties": {
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Fields",
            "type": "object"
          }
        },
        "required": [
          "fields"
        ],
        "title": "UpdatePartFieldsBody",
        "type": "object"
      },
      "UpdatePartPriceBody": {
        "properties": {
          "ext_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ext Price"
          },
          "unit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price"
          }
        },
        "title": "UpdatePartPriceBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "dubIS",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/adjustments/by-source/{source}": {
      "delete": {
        "operationId": "rollback_source",
        "parameters": [
          {
            "in": "path",
            "name": "source",
            "required": true,
            "schema": {
              "title": "Source",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Rollback Source",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rollback Source",
        "tags": [
          "parts"
        ]
      }
    },
    "/v1/adjustments/last": {
      "delete": {
        "operationId": "remove_last_adjustments",
        "parameters": [
          {
            "in": "query",
            "name": "count",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Count",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Remove Last Adjustments",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Last Adjustments",
        "tags": [
          "parts"
        ]
      }
    },
    "/v1/bom/consume": {
      "post": {
        "operationId": "consume_bom",
        "parameters": [
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumeBomBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Consume Bom",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consume Bom",
        "tags": [
          "parts"
        ]
      }
    },
    "/v1/bom/resolve-spec": {
      "post": {
        "operationId": "resolve_bom_spec",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveBomSpecBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Resolve Bom Spec",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Bom Spec",
        "tags": [
          "parts"
        ]
      }
    },
...
    "/v1/parts/fetch-missing-descriptions": {
      "post": {
        "operationId": "fetch_missing_descriptions",
        "parameters": [
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Fetch Missing Descriptions",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch Missing Descriptions",
        "tags": [
          "parts"
        ]
      }
    },
    "/v1/parts/{part_key}": {
      "delete": {
        "operationId": "delete_part",
        "parameters": [
          {
            "in": "path",
            "name": "part_key",
            "required": true,
            "schema": {
              "title": "Part Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Part",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Part",
        "tags": [
          "parts"
        ]
      },
      "patch": {
        "operationId": "update_part_fields",
        "parameters": [
          {
            "in": "path",
            "name": "part_key",
            "required": true,
            "schema": {
              "title": "Part Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePartFieldsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Part Fields",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Part Fields",
        "tags": [
          "parts"
        ]
      }
    },
    "/v1/parts/{part_key}/adjust": {
      "post": {
        "operationId": "adjust_part",
        "parameters": [
          {
            "in": "path",
            "name": "part_key",
            "required": true,
            "schema": {
              "title": "Part Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustPartBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Adjust Part",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Adjust Part",
        "tags": [
          "parts"
        ]
      }
    },
...
    "/v1/parts/{part_key}/fetched-prices": {
      "post": {
        "operationId": "record_fetched_prices",
        "parameters": [
          {
            "in": "path",
            "name": "part_key",
            "required": true,
            "schema": {
              "title": "Part Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordFetchedPricesBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Record Fetched Prices",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Fetched Prices",
        "tags": [
          "parts"
        ]
      }
    },
...
    "/v1/parts/{part_key}/price": {
      "put": {
        "operationId": "update_part_price",
        "parameters": [
          {
            "in": "path",
            "name": "part_key",
            "required": true,
            "schema": {
              "title": "Part Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePartPriceBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Part Price",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Part Price",
        "tags": [
          "parts"
        ]
      }
    },
...
    "/v1/purchases/import": {
      "post": {
        "operationId": "import_purchases",
        "parameters": [
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportPurchasesBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Import Purchases",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Purchases",
        "tags": [
          "parts"
        ]
      }
    },
    "/v1/purchases/last": {
      "delete": {
        "operationId": "remove_last_purchases",
        "parameters": [
          {
            "in": "query",
            "name": "count",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Count",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Remove Last Purchases",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Last Purchases",
        "tags": [
          "parts"
        ]
      }
    },
...
    "/v1/spec/extract": {
      "post": {
        "operationId": "extract_spec_from_value",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractSpecFromValueBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Extract Spec From Value",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extract Spec From Value",
        "tags": [
          "parts"
        ]
      }
    },
...
  }
}